refactor: Add `ProviderModel`
Adds `ProviderModel` as a layer of abstraction between the presentation layer and service layer.
